requirements.txt
pytest~=7.2.1
requests~=2.28.2
obd~=0.7.1
gps~=3.19
scipy
osrm-py